Add Pyside6 support and add tons of new functions
Follow PEP8 variables names

Add BetterLineEdit() that allow digits display and functionnality like a better QSpinBox/QDoubleSpinBox. Support undos, steps, decimals, hold key to change value

Add ExclusiveCheckBox() that allow linked checkbox to always have only one of them selected.

Add CollapsibleGroupBox()

BaseWindow() is now a BaseMixin() allowing for it to be a QWidget or a QMainWindow or else
Add load/save of default settings of the window

LineditSpoiler() is now a Mixin to allow adding it on things other than QLineEdit()
Systray() is now a Mixin()
EditingFinished() is now a Mixin()

Allow any args to be passed to elements in RowLayout()
RowLayout() now uses the BetterLineEdit() as a default LineEdit() instead of a QSpinBox/QDoubleSpinBox
Improve RowLayout() saving/loading of values